Modified the order of the sections and added a data description for the sessions dataset.
I modified the order of the sections to match the sequence outlined in the individual project instructions(still missing Question part, Visualization part and Method&Plan part). I also added a data description for the sessions dataset, including the number of observations and variables, the meaning of each variable, and the potential issues in the data.